Update datamodule to configure datasets as hyperparameters
Also adds a function to recreate a dataset from a given hyperparameter
config as from a training run that produced a given model. This is to be
used by the prediction code to generate a dataset such that congruent
vectorization will occur as done during training.

Updates the latest M2 experiment file as a test-bed.