Fix how tcn test considers total num classes automatically
Some detection inputs have "missing" categories due to their being
skipped during detection inference. This creates a discrepancy between
quantity of listed categories vs category IDs that can be listed.
Instead consider the max category ID + 1, on the assumption that
category IDs start at 0, which is a strong assumption due to considering
category IDs as the index of prediction outputs to which that category
applies.